refactor: move image to module
